simplify page objects using navigation API
