+++ test/page/dvaf1.aggrievedBasic.view.js
@@ -10,9 +10,7 @@ var AggrievedBasicView = function() {
 AggrievedBasicView.prototype = Object.create( DvaForm1Page.prototype, {
 	get: {
 		value: function() {
-			// navigate from home view
-			this.continue();
-			this.continue();
+			this.gotoAggrievedBasicView();
 		}
 	},
